fix(core): prevent potential double-close of File Descriptors (#4874)

+++ benchmarks/src/main/java/org/questdb/PingMain.java
@@ -47,7 +47,7 @@ public class PingMain {
         long bufSize = 1024;
 
         // blocking client for simplicity
-        int fd = Net.socketTcp(true);
+        long fd = Net.socketTcp(true);
         // DNS resolution is provided by the OS
         long inf = Net.getAddrInfo(host, port);
         // attempt to connect
